pyproject.toml
[project]
name = "dgraphai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.7",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "httpx-oauth>=0.16.1",
    "jinja2>=3.1.6",
    "neo4j>=6.1.0",
    "passlib[bcrypt]>=1.7.4",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.24",
    "sqlalchemy>=2.0.49",
    "strawberry-graphql[fastapi]>=0.314.3",
    "uvicorn[standard]>=0.44.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "anyio>=4.0",
    "faker>=30.0",
    "httpx>=0.28",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "respx>=0.22",
]
